test(svelte): move diagnostic-harness into tests/helpers
Cross-feature after #1408 (diagnostics + interaction suites import it).
CONTRIBUTING.md puts shared helpers under tests/helpers/, not a feature folder.
