+++ packages/svelte/tests/helpers/diagnostic-harness.ts
@@ -1,10 +1,13 @@
 /**
- * Shared harness for browser-lane GGPlot-mount diagnostic tests.
+ * Shared harness for browser-lane GGPlot-mount diagnostic tests
+ * (interaction + composition suites). Lives in tests/helpers/ so multiple
+ * feature folders can import it without crossing feature-to-feature paths.
  *
  * Why browser lane only: every advisory delivery path in plot-engine.svelte.ts
  * (deliverAdvisoriesOnce, the config-diagnostics effect) runs inside $effect,
  * which svelte/server render() never executes. Assertions on ondiagnostic
- * payloads placed in *.ssr.test.ts would be vacuous passes — keep them here.
+ * payloads placed in *.ssr.test.ts would be vacuous passes — keep them in
+ * browser-lane suites that import this helper.
  */
 import { expect } from "vitest";
